requirement.txt
numpy
pandas
tensorflow
pyarrow
ystock
scipy
functools
PyWavelets
Sphinx
rinohtype
nbsphinx
plotly >= 4.8.0
pytest-dependency